requirements.txt
matplotlib>=3.8.0
pandas>=2.1.0
a2a-sdk>=0.3.3
mcp>=1.9.4
azure-ai-agents>=1.1.0b1
semantic-kernel>=1.33.0
azure-functions
GitPython==3.1.41
a2a-sdk[http-server]
gradio>=5.33.2
httpx>=0.25.0
python-dotenv>=1.0.0
click>=8.2.1
crewai[tools]>=0.95.0
google-adk>=1.8.0
google-genai>=1.27.0
sse-starlette>=2.3.6
starlette>=0.46.2
uvicorn>=0.35.0
litellm